Fixed warnings and compilation

+++ DXRplayground/Source/DXrenderer/Mesh.cpp
@@ -28,10 +28,10 @@ Mesh::Mesh(RenderContext& ctx, const std::string& path)
     const tinygltf::Scene& scene = model.scenes[model.defaultScene];
     for (int node : scene.nodes)
         ParseModelNodes(model, model.nodes[node]);
-    m_indexCount = m_indices.size();
+    m_indexCount = static_cast<UINT>(m_indices.size());
 
-    m_vertexBuffer = new VertexBuffer(reinterpret_cast<byte*>(m_vertices.data()), sizeof(Vertex) * m_vertices.size(), sizeof(Vertex), ctx.CommandList, ctx.Device);
-    m_indexBuffer = new IndexBuffer(reinterpret_cast<byte*>(m_indices.data()), sizeof(UINT) * m_indices.size(), ctx.CommandList, ctx.Device, DXGI_FORMAT_R32_UINT);
+    m_vertexBuffer = new VertexBuffer(reinterpret_cast<byte*>(m_vertices.data()), static_cast<UINT>(sizeof(Vertex) * m_vertices.size()), sizeof(Vertex), ctx.CommandList, ctx.Device);
+    m_indexBuffer = new IndexBuffer(reinterpret_cast<byte*>(m_indices.data()), static_cast<UINT>(sizeof(UINT) * m_indices.size()), ctx.CommandList, ctx.Device, DXGI_FORMAT_R32_UINT);
 }
 
 Mesh::Mesh(RenderContext& ctx, std::vector<Vertex> vertices, std::vector<UINT> indices)
@@ -39,8 +39,8 @@ Mesh::Mesh(RenderContext& ctx, std::vector<Vertex> vertices, std::vector<UINT> i
     m_vertices.swap(vertices);
     m_indices.swap(indices);
 
-    m_vertexBuffer = new VertexBuffer(reinterpret_cast<byte*>(m_vertices.data()), sizeof(Vertex) * m_vertices.size(), sizeof(Vertex), ctx.CommandList, ctx.Device);
-    m_indexBuffer = new IndexBuffer(reinterpret_cast<byte*>(m_indices.data()), sizeof(UINT) * m_indices.size(), ctx.CommandList, ctx.Device, DXGI_FORMAT_R32_UINT);
+    m_vertexBuffer = new VertexBuffer(reinterpret_cast<byte*>(m_vertices.data()), static_cast<UINT>(sizeof(Vertex) * m_vertices.size()), sizeof(Vertex), ctx.CommandList, ctx.Device);
+    m_indexBuffer = new IndexBuffer(reinterpret_cast<byte*>(m_indices.data()), static_cast<UINT>(sizeof(UINT) * m_indices.size()), ctx.CommandList, ctx.Device, DXGI_FORMAT_R32_UINT);
 }
 
 Mesh::~Mesh()
